move neo4j-inserter back to parent directory
